select closest veturilo point

+++ app/src/main/kotlin/com/legec/ceburilo/web/veturilo/VeturiloCountry.kt
@@ -7,9 +7,9 @@ import org.simpleframework.xml.Root
 @Root(name = "country")
 class VeturiloCountry {
     @field:Attribute(name = "lat")
-    var latitude: Float = 0f
+    var latitude: Double = 0.0
     @field:Attribute(name = "lng")
-    var longtitude: Float = 0f
+    var longtitude: Double = 0.0
     @field:Attribute
     var zoom: Int = 0
     @field:Attribute
